+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # CHANGELOG
 
 ## 1.16.0
-### 06-Feb-2025
+### 18-Feb-2025
 - Added proof-of-concept validations for swiftDialog `2.5.1`'s "blurscreen" control
 - Removed vendor-specific Local Validations (in favor of Remote Validations)
 - Updated Configuration `policyJSON` to better match internal usage
 - Added "activate" command to Validations
 - Fixed minor issue with `calculateFreeDiskSpace` function result not being parsed into scriptLog (thanks, @robjschroeder!)
+- Updated the MS Teams message template to the new format #156 (thanks, @nlopezUA!)
 
 ## 1.15.1
 ### 06-Feb-2025